split info to files
