+++ src/constants/networks/arbitrum-testnet.ts
@@ -1,0 +1,55 @@
+import { ChainId } from '@kyberswap/ks-sdk-core'
+import { createClient, NetworkInfo } from 'constants/networks'
+
+import ARBITRUM from '../assets/networks/arbitrum-network.svg'
+import EthereumLogo from '../assets/images/ethereum-logo.png'
+
+const EMPTY = ''
+const EMPTY_ARRAY: any[] = []
+
+const arbitrumTestnetInfo: NetworkInfo = {
+  chainId: ChainId.ARBITRUM_TESTNET,
+  route: 'arbitrum-testnet',
+  name: 'Arbitrum Testnet',
+  icon: ARBITRUM,
+  classicClient: createClient('https://api.thegraph.com/subgraphs/name/viet-nv/kyberswap-arbitrum-rinkeby'),
+  elasticClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/kyberswap-elastic-arbitrum-one'),
+  blockClient: createClient('https://api.thegraph.com/subgraphs/name/viet-nv/arbitrum-rinkeby-blocks'),
+  etherscanUrl: 'https://testnet.arbiscan.io',
+  etherscanName: 'Arbiscan',
+  tokenListUrl: 'https://raw.githubusercontent.com/KyberNetwork/ks-assets/main/tokenLists/arbitrum.tokenlist.json',
+  bridgeURL: 'https://bridge.arbitrum.io',
+  nativeToken: {
+    symbol: 'ETH',
+    name: 'ETH (Wrapped)',
+    address: '0xB47e6A5f8b33b3F17603C83a0535A9dcD7E32681',
+    logo: EthereumLogo,
+    decimal: 18,
+  },
+  rpcUrl: 'https://rinkeby.arbitrum.io/rpc',
+  routerUri: EMPTY,
+  classic: {
+    zap: '0xfa33723F6fA00a35F69F8aCd72A5BE9AF3c8Bd25',
+    router: '0x78Ad9A49327D73C6E3B9881eCD653232cF3E480C',
+    routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
+    aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
+    factory: '0x9D4ffbf49cc21372c2115Ae4C155a1e5c0aACf36',
+    migrate: EMPTY,
+    claimReward: EMPTY,
+    fairlaunch: EMPTY_ARRAY,
+    fairlaunchV2: EMPTY_ARRAY,
+  },
+  elastic: {
+    coreFactory: '0x5F1dddbf348aC2fbe22a163e30F99F9ECE3DD50a',
+    nonfungiblePositionManager: '0x2B1c7b41f6A8F2b2bc45C3233a5d5FB3cD6dC9A8',
+    tickReader: '0xe3AC3fd66EB31cAf4EE0831b262D837c479FFCe5',
+    initCodeHash: '0xc597aba1bb02db42ba24a8878837965718c032f8b46be94a6e46452a9f89ca01',
+    quoter: '0x0D125c15D54cA1F8a813C74A81aEe34ebB508C1f',
+    routers: '0xC1e7dFE73E1598E3910EF4C7845B68A9Ab6F4c83',
+  },
+  avgrageBlockTimeInSeconds: 1, // TODO: check these info
+  coingeckoNetworkId: EMPTY,
+  coingeckoNativeTokenId: 'ethereum',
+}
+
+export default arbitrumTestnetInfo
