+++ src/constants/networks/avax.ts
@@ -1,0 +1,63 @@
+import { ChainId } from '@kyberswap/ks-sdk-core'
+import { createClient, NetworkInfo } from 'constants/networks'
+
+import AVAX from '../assets/networks/avax-network.png'
+
+const EMPTY = ''
+
+const avaxInfo: NetworkInfo = {
+  chainId: ChainId.AVAXMAINNET,
+  route: 'avalanche',
+  name: 'Avalanche',
+  icon: AVAX,
+  classicClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/kyberswap-exchange-avalanche'),
+  elasticClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/kyberswap-elastic-avalanche'),
+  blockClient: createClient('https://api.thegraph.com/subgraphs/name/ducquangkstn/avalache-blocks'),
+  etherscanUrl: 'https://snowtrace.io',
+  etherscanName: 'Snowtrace',
+  tokenListUrl: 'https://raw.githubusercontent.com/KyberNetwork/ks-assets/main/tokenLists/avax.mainnet.tokenlist.json',
+  bridgeURL: 'https://bridge.avax.network',
+  nativeToken: {
+    symbol: 'AVAX',
+    name: 'AVAX (Wrapped)',
+    address: '0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7',
+    logo: AVAX,
+    decimal: 18,
+  },
+  rpcUrl: 'https://avalanche.dmm.exchange/v1/mainnet/geth?appId=prod-dmm',
+  routerUri: `${process.env.REACT_APP_AGGREGATOR_API}/avalanche/route/encode`,
+  classic: {
+    zap: '0x83D4908c1B4F9Ca423BEE264163BC1d50F251c31',
+    router: '0x8Efa5A9AD6D594Cf76830267077B78cE0Bc5A5F8',
+    routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
+    aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
+    factory: '0x10908C875D865C66f271F5d3949848971c9595C9',
+    migrate: EMPTY,
+    claimReward: '0x610A05127d51dd42031A39c25aF951a8e77cDDf7',
+    fairlaunch: [
+      '0xD169410524Ab1c3C51F56a856a2157B88d4D4FF5',
+      '0x3133C5C35947dBcA7A76Ee05f106a7c63BFD5C3F',
+      '0x98910F7f13496fcDE2ade93648F05b4854Fc99D9',
+      '0x854Cf246b09c7366AEe5abce92fA167bfE7f3E75',
+    ],
+    fairlaunchV2: [
+      '0x8e9Bd30D15420bAe4B7EC0aC014B7ECeE864373C',
+      '0x845d1d0d9b344fba8a205461b9e94aefe258b918',
+      '0xa107e6466Be74361840059a11e390200371a7538',
+      '0x89929Bc485cE72D2Af7b7283B40b921e9F4f80b3',
+    ],
+  },
+  elastic: {
+    coreFactory: '0x5F1dddbf348aC2fbe22a163e30F99F9ECE3DD50a',
+    nonfungiblePositionManager: '0x2B1c7b41f6A8F2b2bc45C3233a5d5FB3cD6dC9A8',
+    tickReader: '0xe3AC3fd66EB31cAf4EE0831b262D837c479FFCe5',
+    initCodeHash: '0xc597aba1bb02db42ba24a8878837965718c032f8b46be94a6e46452a9f89ca01',
+    quoter: '0x0D125c15D54cA1F8a813C74A81aEe34ebB508C1f',
+    routers: '0xC1e7dFE73E1598E3910EF4C7845B68A9Ab6F4c83',
+  },
+  avgrageBlockTimeInSeconds: 1.85,
+  coingeckoNetworkId: 'avalanche',
+  coingeckoNativeTokenId: 'avalanche-2',
+}
+
+export default avaxInfo
