+++ src/constants/networks/bttc.ts
@@ -1,0 +1,62 @@
+import { ChainId } from '@kyberswap/ks-sdk-core'
+import { createClient, NetworkInfo } from 'constants/networks'
+
+import BTT from '../assets/networks/bttc.png'
+
+const EMPTY = ''
+const EMPTY_ARRAY: any[] = []
+
+const bttcInfo: NetworkInfo = {
+  chainId: ChainId.BTTC,
+  route: 'bittorrent',
+  name: 'BitTorrent',
+  icon: BTT,
+  classicClient: createClient(
+    'https://bttc-graph.kyberengineering.io/subgraphs/name/kybernetwork/kyberswap-exchange-bttc',
+  ),
+  elasticClient: createClient(
+    'https://bttc-graph.kyberengineering.io/subgraphs/name/kybernetwork/kyberswap-elastic-bttc',
+  ),
+  blockClient: createClient('https://bttc-graph.kyberengineering.io/subgraphs/name/kybernetwork/bttc-blocks'),
+  etherscanUrl: 'https://bttcscan.com',
+  etherscanName: 'Bttcscan',
+  tokenListUrl: 'https://raw.githubusercontent.com/KyberNetwork/ks-assets/main/tokenLists/bttc.tokenlist.json',
+  bridgeURL: 'https://wallet.bt.io/bridge',
+  nativeToken: {
+    symbol: 'BTT',
+    name: 'BTT (Wrapped)',
+    address: '0x8D193c6efa90BCFf940A98785d1Ce9D093d3DC8A',
+    logo: BTT,
+    decimal: 18,
+  },
+  rpcUrl: 'https://rpc.bt.io',
+  routerUri: `${process.env.REACT_APP_AGGREGATOR_API}/bttc/route/encode`,
+  classic: {
+    zap: '0x83D4908c1B4F9Ca423BEE264163BC1d50F251c31',
+    router: '0xEaE47c5D99f7B31165a7f0c5f7E0D6afA25CFd55',
+    routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
+    aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
+    factory: '0xD9bfE9979e9CA4b2fe84bA5d4Cf963bBcB376974',
+    migrate: EMPTY,
+    claimReward: '0x1a91f5ADc7cB5763d35A26e98A18520CB9b67e70',
+    fairlaunch: EMPTY_ARRAY,
+    fairlaunchV2: [
+      '0x8e9Bd30D15420bAe4B7EC0aC014B7ECeE864373C',
+      '0xa107e6466Be74361840059a11e390200371a7538',
+      '0x89929Bc485cE72D2Af7b7283B40b921e9F4f80b3',
+    ],
+  },
+  elastic: {
+    coreFactory: '0x5F1dddbf348aC2fbe22a163e30F99F9ECE3DD50a',
+    nonfungiblePositionManager: '0x2B1c7b41f6A8F2b2bc45C3233a5d5FB3cD6dC9A8',
+    tickReader: '0xe3AC3fd66EB31cAf4EE0831b262D837c479FFCe5',
+    initCodeHash: '0xc597aba1bb02db42ba24a8878837965718c032f8b46be94a6e46452a9f89ca01',
+    quoter: '0x0D125c15D54cA1F8a813C74A81aEe34ebB508C1f',
+    routers: '0xC1e7dFE73E1598E3910EF4C7845B68A9Ab6F4c83',
+  },
+  avgrageBlockTimeInSeconds: 2, // TODO: check these info
+  coingeckoNetworkId: 'tron',
+  coingeckoNativeTokenId: 'bittorrent',
+}
+
+export default bttcInfo
