+++ src/constants/networks/fantom.ts
@@ -1,0 +1,55 @@
+import { ChainId } from '@kyberswap/ks-sdk-core'
+import { createClient, NetworkInfo } from 'constants/networks'
+
+import FTM from '../assets/networks/fantom-network.png'
+
+const EMPTY = ''
+const EMPTY_ARRAY: any[] = []
+
+const fantomInfo: NetworkInfo = {
+  chainId: ChainId.FANTOM,
+  route: 'fantom',
+  name: 'Fantom',
+  icon: FTM,
+  classicClient: createClient('https://api.thegraph.com/subgraphs/name/dynamic-amm/dmm-exchange-ftm'),
+  elasticClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/kyberswap-elastic-fantom'),
+  blockClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/fantom-blocks'),
+  etherscanUrl: 'https://ftmscan.com',
+  etherscanName: 'Ftmscan',
+  tokenListUrl:
+    'https://raw.githubusercontent.com/KyberNetwork/ks-assets/main/tokenLists/fantom.mainnet.tokenlist.json',
+  bridgeURL: 'https://multichain.xyz',
+  nativeToken: {
+    symbol: 'FTM',
+    name: 'FTM (Wrapped)',
+    address: '0x21be370D5312f44cB42ce377BC9b8a0cEF1A4C83',
+    logo: FTM,
+    decimal: 18,
+  },
+  rpcUrl: 'https://rpc.ftm.tools',
+  routerUri: `${process.env.REACT_APP_AGGREGATOR_API}/fantom/route/encode`,
+  classic: {
+    zap: '0x83D4908c1B4F9Ca423BEE264163BC1d50F251c31',
+    router: '0x5d5A5a0a465129848c2549669e12cDC2f8DE039A',
+    routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
+    aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
+    factory: '0x78df70615ffc8066cc0887917f2Cd72092C86409',
+    migrate: EMPTY,
+    claimReward: EMPTY,
+    fairlaunch: EMPTY_ARRAY,
+    fairlaunchV2: EMPTY_ARRAY,
+  },
+  elastic: {
+    coreFactory: '0x5F1dddbf348aC2fbe22a163e30F99F9ECE3DD50a',
+    nonfungiblePositionManager: '0x2B1c7b41f6A8F2b2bc45C3233a5d5FB3cD6dC9A8',
+    tickReader: '0xe3AC3fd66EB31cAf4EE0831b262D837c479FFCe5',
+    initCodeHash: '0xc597aba1bb02db42ba24a8878837965718c032f8b46be94a6e46452a9f89ca01',
+    quoter: '0x0D125c15D54cA1F8a813C74A81aEe34ebB508C1f',
+    routers: '0xC1e7dFE73E1598E3910EF4C7845B68A9Ab6F4c83',
+  },
+  avgrageBlockTimeInSeconds: 1,
+  coingeckoNetworkId: 'fantom',
+  coingeckoNativeTokenId: 'fantom',
+}
+
+export default fantomInfo
