+++ src/constants/networks/matic.ts
@@ -1,0 +1,61 @@
+import { ChainId } from '@kyberswap/ks-sdk-core'
+import { createClient, NetworkInfo } from 'constants/networks'
+
+import Polygon from '../assets/networks/polygon-network.png'
+
+const EMPTY = ''
+const EMPTY_ARRAY: any[] = []
+
+const maticInfo: NetworkInfo = {
+  chainId: ChainId.MATIC,
+  route: 'polygon',
+  name: 'Polygon',
+  icon: Polygon,
+  classicClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/kyberswap-exchange-polygon'),
+  elasticClient: createClient('https://api.thegraph.com/subgraphs/name/kybernetwork/kyberswap-elastic-matic'),
+  blockClient: createClient('https://api.thegraph.com/subgraphs/name/dynamic-amm/ethereum-blocks-polygon'),
+  etherscanUrl: 'https://polygonscan.com',
+  etherscanName: 'Polygonscan',
+  tokenListUrl: 'https://raw.githubusercontent.com/KyberNetwork/ks-assets/main/tokenLists/matic.tokenlist.json',
+  bridgeURL: 'https://wallet.matic.network/bridge',
+  nativeToken: {
+    symbol: 'MATIC',
+    name: 'MATIC (Wrapped)',
+    address: '0x0d500B1d8E8eF31E21C99d1Db9A6444d3ADf1270',
+    logo: Polygon,
+    decimal: 18,
+  },
+  rpcUrl: 'https://polygon.dmm.exchange/v1/mainnet/geth?appId=prod-dmm',
+  routerUri: EMPTY,
+  classic: {
+    zap: '0x83D4908c1B4F9Ca423BEE264163BC1d50F251c31',
+    router: '0x546C79662E028B661dFB4767664d0273184E4dD1',
+    routerV2: '0x00555513Acf282B42882420E5e5bA87b44D8fA6E',
+    aggregationExecutor: '0x41684b361557E9282E0373CA51260D9331e518C9',
+    factory: '0x5F1fe642060B5B9658C15721Ea22E982643c095c',
+    migrate: EMPTY,
+    claimReward: '0x89929Bc485cE72D2Af7b7283B40b921e9F4f80b3',
+    fairlaunch: [
+      '0xc39bD0fAE646Cb026C73943C5B50E703de2a6532',
+      '0xc940acee228893c14274eF1bB64e631308E96e1A',
+      '0x7EB05d3115984547a50Ff0e2d247fB6948E1c252',
+      '0xc0601973451d9369252Aee01397c0270CD2Ecd60',
+      '0x829c27fd3013b944cbE76E92c3D6c45767c0C789',
+      '0x3aDd3034Fcf921F20c74c6149FB44921709595B1',
+    ],
+    fairlaunchV2: EMPTY_ARRAY,
+  },
+  elastic: {
+    coreFactory: '0x5F1dddbf348aC2fbe22a163e30F99F9ECE3DD50a',
+    nonfungiblePositionManager: '0x2B1c7b41f6A8F2b2bc45C3233a5d5FB3cD6dC9A8',
+    tickReader: '0xe3AC3fd66EB31cAf4EE0831b262D837c479FFCe5',
+    initCodeHash: '0xc597aba1bb02db42ba24a8878837965718c032f8b46be94a6e46452a9f89ca01',
+    quoter: '0x0D125c15D54cA1F8a813C74A81aEe34ebB508C1f',
+    routers: '0xC1e7dFE73E1598E3910EF4C7845B68A9Ab6F4c83',
+  },
+  avgrageBlockTimeInSeconds: 2.6,
+  coingeckoNetworkId: 'polygon-pos',
+  coingeckoNativeTokenId: 'matic-network',
+}
+
+export default maticInfo
